organize the parser a little better
also worth noting: the reason that Term now just contains a reference
to the string literal instead of an actual instance of the appropriate
type is that it makes things much simpler: we don't have to back-propagate
the actual desired type all the way to the parse layer, we can just
figure out what type is desired later. Also, string references are
Clone+Copy+Eq, which not all numeric types are.