remove custom output formatting. write type trait. impl for f64.
Custom output formatting turned out to be a huge can of worms which
just wasn't worth it. Doing it right is going to end up being a big
project of its own. The best approach so far is likely going to be
extending the num_formatting crate with i.e. custom grouping spacings,
but for now, it's just not worth it.

See: https://crates.io/crates/num-format

Removing that requirement means that we can now extract into a trait
all the things that this calculator can do. Implemented that trait
for f64. In principle, we can continue now.