container actions are only available on linux
This was breaking the builds which attempted to publish the
build artifacts in the same job that built them.

Instead, we have to introduce a dependent job which publishes
artifacts to the appropriate release after the artifacts have all
been built.

Also, we add caching, because there's no reason not to, and it
could speed things up.