Don't evaluate CLI expressions twice
Fixes #14.

We'd been evaluating expressions twice: once to get a numeric result
to insert into the history, and again to get the string output which
was formatted according to user request. That turned out to be a bad
idea when the expression contained a back-reference, because the
back-reference was updated before the display version was re-calculated.

Evaluating once and then updating history is absolutely the way to
go here.